Added new file VirtualListNative.js and added code removed from VirtualList to VirtualListNative
